Change sequence length type to int

+++ estimator/turbofan/src/main/kotlin/com/alessandropacielli/turbofan/regression/tensorflow/serving/HttpRnnEstimatorProxy.java
@@ -3,20 +3,17 @@ package com.alessandropacielli.turbofan.regression.tensorflow.serving;
 import com.alessandropacielli.turbofan.regression.RemainingLifeEstimator;
 import org.jetbrains.annotations.NotNull;
 import org.nd4j.linalg.api.ndarray.INDArray;
-import org.nd4j.linalg.cpu.nativecpu.NDArray;
 import org.nd4j.linalg.factory.Nd4j;
 import org.springframework.http.HttpEntity;
 import org.springframework.web.client.RestTemplate;
-
-import java.util.List;
 
 public class HttpRnnEstimatorProxy implements RemainingLifeEstimator {
 
     private String url;
     private RestTemplate restTemplate;
-    private Long sequenceLength;
+    private int sequenceLength;
 
-    public HttpRnnEstimatorProxy(String url, Long sequenceLenght, RestTemplate restTemplate) {
+    public HttpRnnEstimatorProxy(String url, int sequenceLenght, RestTemplate restTemplate) {
         this.url = url;
         this.restTemplate = restTemplate;
         this.sequenceLength = sequenceLenght;
@@ -33,13 +30,13 @@ public class HttpRnnEstimatorProxy implements RemainingLifeEstimator {
     }
 
     @Override
-    public long getSequenceLength() {
+    public int getSequenceLength() {
         return this.sequenceLength;
     }
 
     public static void main(String[] args) {
         RestTemplate restTemplate = new RestTemplate();
-        RemainingLifeEstimator rulRegressor = new HttpRnnEstimatorProxy("http://localhost:8501/v1/models/turbofan:predict", 50L, restTemplate);
+        RemainingLifeEstimator rulRegressor = new HttpRnnEstimatorProxy("http://localhost:8501/v1/models/turbofan:predict", 50, restTemplate);
         INDArray ndArray = Nd4j.zeros( 50, 25);
         System.out.println(rulRegressor.predict(ndArray.toDoubleMatrix()));
     }
